Use django-social-auth to log users in with Google
You need to get the client ID and SECRET from Google.
Go to https://code.google.com/apis/console/ and create a project.
On that project, click on 'API Access' and create an OAuth 2 Client ID.
Choose a 'Web application' as the type, click 'more options' and enter http://localhost:8000/complete/google-oauth2/ as the redirect URI.

This will work for testing and development, but you need to enter the real domain for your project.